fix: header files not being found on linux
Lowercase the names of imported header files to fix a problem with header files not being found on linux.

when compile via cargo-xwin，it cannot found header file beacause of Uppercase name

This commit will not affect compilation on windows because windows is case insensitive when looking for header files.