fix: cross compiling windows on linux case sensitivity
(Explainer on original Pull Request):
Lowercase the names of imported header files to fix a problem with header files not being found on Linux.
When compiling via cargo-xwin，Synchapi.h / Threadpoolapiset.h cannot be found beacause of the Uppercased first letter.
This commit will not affect compilation on Windows because it is case insensitive when looking for header files.

Co-authored-by: Akkuman <akkumans@qq.com>